Move RSpec helpers to their own file
